+++ src/bin/mie-decoder.rs
@@ -1,3 +1,5 @@
+#![cfg_attr(not(test), warn(clippy::expect_used, clippy::unwrap_used))]
+
 use std::process::ExitCode;
 
 fn main() -> ExitCode {